+++ source/extensions/lightspeed.trex.utils.widget/lightspeed/trex/utils/widget/tests/__init__.py
@@ -17,6 +17,7 @@
 
 from .e2e.test_invalid_deps_dialog import TestInvalidDepsDialog
 from .unit.test_ingestcraft_loader import TestIngestCraftLoader
+from .unit.test_message_dialog import TestMessageDialog
 from .unit.test_quicklayout import TestQuickLayout
 
-__all__ = ["TestIngestCraftLoader", "TestInvalidDepsDialog", "TestQuickLayout"]
+__all__ = ["TestIngestCraftLoader", "TestInvalidDepsDialog", "TestMessageDialog", "TestQuickLayout"]